+++ lib/algae/tree/binary_search/chain.ex
@@ -1,0 +1,14 @@
+alias Algae.Tree.BinarySearch, as: BST
+alias Algae.Tree.BinarySearch.{Empty, Node}
+
+import TypeClass
+
+use Witchcraft
+
+definst Witchcraft.Chain, for: Algae.Tree.BinarySearch.Empty do
+  def chain(_, _), do: %Empty{}
+end
+
+definst Witchcraft.Chain, for: Algae.Tree.BinarySearch.Node do
+  def chain(%Node{node: node}, link), do: link.(node)
+end
